Factored out generic response exception
